Refactor and restructure valuation, add valuation README
