+++ example/valuation/_util.py
@@ -1,0 +1,22 @@
+import torch
+
+
+def fuzzy_position(pos1, pos2, keyword):
+    x = pos2[:, 0] - pos1[:, 0]
+    y = pos2[:, 1] - pos1[:, 1]
+    tan = torch.atan2(y, x)
+    degree = tan[:] / torch.pi * 180
+    if keyword == 'top':
+        probs = 1 - abs(degree[:] - 90) / 90
+        result = torch.where((180 >= degree) & (degree >= 0), probs * 0.9, 0.)
+    elif keyword == 'left':
+        probs = (abs(degree[:]) - 90) / 90
+        result = torch.where((degree <= -90) | (degree >= 90), probs * 0.9, 0.)
+    elif keyword == 'bottom':
+        probs = 1 - abs(degree[:] + 90) / 90
+        result = torch.where((0 >= degree) & (degree >= -180), probs * 0.9, 0.)
+    elif keyword == 'right':
+        probs = 1 - abs(degree[:]) / 90
+        result = torch.where((90 >= degree) & (degree >= -90), probs * 0.9, 0.)
+
+    return result
